Menu2

+++ team_jg_entertainment/src/main/java/lv/javaguru/java2/jg_entertainment/restaurant/core/requests/menus/AddMenuRequest.java
@@ -6,7 +6,7 @@ public class AddMenuRequest {
     private String description;
     private double price;
 
-    public AddMenuRequest(String title, String description, int price) {
+    public AddMenuRequest(String title, String description, double price) {
         this.title = title;
         this.description = description;
         this.price = price;
